orb-discovery/orb-discovery/pyproject.toml
[project]
name = "orb-discovery"
version = "0.0.1"  # Overwritten during the build process
description = "NetBox Labs, Orb Discovery"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "Apache" }
authors = [
    {name = "NetBox Labs", email = "support@netboxlabs.com" }
]
maintainers = [
    {name = "NetBox Labs", email = "support@netboxlabs.com" }
]

classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Build Tools",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    'Programming Language :: Python :: 3.10',
    'Programming Language :: Python :: 3.11',
    'Programming Language :: Python :: 3.12',
]

dependencies = [
    "APScheduler~=3.10",
    "croniter~=5.0",
    "fastapi~=0.115",
    "httpx~=0.27",
    "importlib-metadata~=8.5",
    "napalm~=5.0",
    "netboxlabs-diode-sdk~=0.4",
    "pydantic~=2.9",
    "python-dotenv~=1.0",
    "uvicorn~=0.32",
]

[project.optional-dependencies]
dev = ["black", "check-manifest", "ruff"]
test = ["coverage", "pytest", "pytest-cov"]

[project.urls]
"Homepage" = "https://netboxlabs.com/"

[project.scripts]
orb-discovery = "orb_discovery.main:main"

[tool.setuptools]
packages = [
    "orb_discovery",
    "orb_discovery.policy",
]

[build-system]
requires = ["setuptools>=43.0.0", "wheel"]
build-backend = "setuptools.build_meta"


[tool.ruff]
line-length = 140

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.ruff.lint]
select = ["C", "D", "E", "F", "I", "R", "UP", "W"]
ignore = ["F401", "D203", "D212", "D400", "D401", "D404", "RET504"]